% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ggiso.R
\name{auto_ggiso}
\alias{auto_ggiso}
\title{Make isotope plot.}
\usage{
auto_ggiso(
  .data,
  .mapp = aes(...),
  .xlim,
  .ylim,
  .shape_val = c(16, 17, 24),
  .xlab,
  .ylab,
  .height = 0.2,
  .width = 0.2,
  .hjust = 0,
  .vjust = 0,
  .breaks,
  .size = 0.7,
  .stroke = 0.9,
  .point_size = 2,
  .auto_fix = T,
  .ratio
)
}
\arguments{
\item{.data}{data}

\item{.xlim}{range of x axis.}

\item{.ylim}{range of y axis.}

\item{.shape_val}{shape of point, passing scale_shape_manual().}

\item{.xlab}{title of x axis.}

\item{.ylab}{title of y axis.}

\item{.height}{height of errorbar.}

\item{.width}{width of errorbar.}

\item{.hjust}{place of label (horizontal).}

\item{.vjust}{place of label (vertical).}

\item{.breaks}{number of breaks.}

\item{.size}{width of line in errorbar.}

\item{.stroke}{edge of point}

\item{.point_size}{size of point.}

\item{.auto_fix}{fix ratio of plot automatically.}

\item{.ratio}{ratio of plot.}

\item{.mapping}{mapping, wrapping by aes(), augments passing through geom_***.}
}
\description{
Make isotope plot with errobar automatically. If you regulate more options, use ggiso().
}
