% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocat_plot.R
\name{cocat_plot}
\alias{cocat_plot}
\title{Concatenate plots.}
\usage{
cocat_plot(.obj)
}
\arguments{
\item{.obj}{data.frame}
}
\description{
Concatenate plots in one column. This makes more powerful by using with purrr::map function.
}
