% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixlimit.R
\name{fixlmit}
\alias{fixlmit}
\title{fixlmit}
\usage{
fixlmit(.plot, .xlim, .ylim, .ratio)
}
\arguments{
\item{.plot}{ggplot object.}

\item{.xlim}{xlim}

\item{.ylim}{ylim}

\item{.ratio}{aspect ratio}
}
\description{
fixlmit
}
