% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcairo.R
\name{ggcairo}
\alias{ggcairo}
\title{Save plot by useing cairo_pdf}
\usage{
ggcairo(.path, .plot = ggplot2::last_plot(), .family = "Arial", ...)
}
\arguments{
\item{.path}{path}

\item{.plot}{ggplot object}

\item{.family}{character, font family.}

\item{...}{other augument pa}
}
\description{
Save plot by useing cairo_pdf. By default, saving last plot.
}
