% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_str.R
\name{lm_str}
\alias{lm_str}
\title{Make simple lm model equation.}
\usage{
lm_str(.resp = "y", .intrspt = NULL, .expl = NULL, .parse = T, .digits = 2)
}
\arguments{
\item{.resp}{character, respose variable.}

\item{.intrspt}{intersept}

\item{.expl}{explanatory variable.}

\item{.parse}{parse}

\item{.digit}{digit}
}
\description{
Make simple lm model equation as character.
}
