% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ggenrich.R
\name{auto_ggenrich}
\alias{auto_ggenrich}
\title{Make enrichment factor plot.}
\usage{
auto_ggenrich(
  .data,
  .data_ref,
  .mapping = aes(...),
  .mapping_ref = aes(...),
  .xlim,
  .ylim,
  .xlab,
  .ylab,
  .point_size,
  .stroke,
  .shape_val = c(16, 17, 24),
  .width = 0.2,
  .height = 0.2,
  .linesize,
  .linetype,
  .hjust = 0,
  .vjust = 0,
  .breaks,
  .auto_fix = T,
  .ratio,
  ...
)
}
\arguments{
\item{.data}{data.}

\item{.data_ref}{data of reference autotrophic plants.}

\item{.mapping}{mapping, wrapping by aes(), augments passing through geom_***.}

\item{.mapping_ref}{mapping, reference data set.}

\item{.xlim}{range of x axis.}

\item{.ylim}{range of y axis.}

\item{.xlab}{title of x axis.}

\item{.ylab}{title of y axis.}

\item{.point_size}{size of point.}

\item{.stroke}{edge of point}

\item{.shape_val}{shape of point, passing scale_shape_manual().}

\item{.width}{width of error bar.}

\item{.height}{height of error bar.}

\item{.linesize}{width of line in error bar.}

\item{.linetype}{dashed or dotted and so on.}

\item{.hjust}{place of label (horizontal).}

\item{.vjust}{place of label (vertical).}

\item{.breaks}{number of breaks.}

\item{.auto_fix}{fix ratio of plot automatically.}

\item{.ratio}{ratio of plot.}
}
\description{
Make enrichment factor plot. If you regulate more options, use ggenrich().
}
