% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ggiso.R
\name{auto_ggiso}
\alias{auto_ggiso}
\title{Make isotope plot.}
\usage{
auto_ggiso(
  .data,
  .mapping,
  .xlim,
  .ylim,
  .shape_val = c(16, 17, 24),
  .xlab = expression(paste(italic("δ"^{     13 }), "C", " (‰)")),
  .ylab = expression(paste(italic("δ"^{     15 }), "N", " (‰)")),
  .cross_tip = 0.15,
  .hjust = 0,
  .vjust = 0,
  .x_breaks = scales::breaks_extended(6),
  .y_breaks = scales::breaks_extended(6),
  .stroke = 0.82,
  .point_size = 3,
  .linesize = 0.55,
  .axis_size = 0.55,
  .family = "Arial",
  .lab_title_size = 15,
  .lab_text_size = 12,
  .label_size = 4,
  .ratio = 2/3,
  .multi = F,
  .scaling = F,
  .scale_var = NULL,
  ...
)
}
\arguments{
\item{.data}{data}

\item{.mapping}{mapping, wrapping by aes(), augments passing through geom_***.}

\item{.xlim}{range of x axis.}

\item{.ylim}{range of y axis.}

\item{.shape_val}{shape of point, passing scale_shape_manual().}

\item{.xlab}{title of x axis.}

\item{.ylab}{title of y axis.}

\item{.cross_tip}{width and height of errorbar.}

\item{.hjust}{place of label (horizontal).}

\item{.vjust}{place of label (vertical).}

\item{.x_breaks}{break of x.}

\item{.y_breaks}{break of y.}

\item{.stroke}{edge of point}

\item{.point_size}{size of point.}

\item{.linesize}{width of line in errorbar.}

\item{.family}{font family, default Arial.}

\item{.lab_title_size}{axis title.}

\item{.lab_text_size}{axis text.}

\item{.label_size}{label size.}

\item{.ratio}{ratio of plot.}

\item{.multi}{logical}

\item{.scaling}{logical}

\item{.scale_var}{numeric}

\item{...}{passed through theme().}
}
\description{
Make isotope plot with errobar automatically. If you regulate more options, use ggiso().
}
\examples{
library(ggplot2)
library(dplyr)

data("demo")

by_place <- demo \%>\%
  group_by(place, label) \%>\%
  summarise(across(where(is.numeric),
                   list(mean = mean, sd = sd),
                   .names = "{.fn}_{.col}"))

place_A <- by_place \%>\%
  filter(place == "A")

# make plot (default).
auto_ggiso(.data = place_A,
           .mapping = aes(x = mean_delta13C, y = mean_delta15N,
                          xmin = mean_delta13C - sd_delta13C,
                          xmax = mean_delta13C + sd_delta13C,
                          ymin = mean_delta15N - sd_delta15N,
                          ymax = mean_delta15N + sd_delta15N,
                          shape = label,
                          label = label),
          .xlim = c(-36, -26),
          .ylim = c(-6, 4),
          .family = NULL,
          .x_breaks = seq(-36, -26, 2),
          .y_breaks = seq(-6, 4, 2))

# make plot (manual).
auto_ggiso(.data = place_A,
           .mapping = aes(x = mean_delta13C, y = mean_delta15N,
                          xmin = mean_delta13C - sd_delta13C,
                          xmax = mean_delta13C + sd_delta13C,
                          ymin = mean_delta15N - sd_delta15N,
                          ymax = mean_delta15N + sd_delta15N,
                          shape = label,
                          label = label),
          .xlim = c(-36, -26),
          .ylim = c(-6, 4),
          .xlab = expression(paste(italic("δ"^{13}), "C", " (\u2030)")),
          .ylab = expression(paste(italic("δ"^{15}), "N", " (\u2030)")),
          .hjust = 0,
          .vjust = 0,
          .cross_tip = .3,
          .shape_val = c(24, 21, 16, 23),
          .stroke = 1,
          .axis_size = .8,
          .linesize = .8,
          .x_breaks = scales::breaks_extended(6),
          .y_breaks = scales::breaks_extended(6),
          .point_size = 3,
          .family = NULL,
          .label_size = 4,
          .lab_title_size = 15,
          .lab_text_size = 12,
          .ratio = 1/2)
}
